reorg files to get tests passing
