+++ src/app/backend/code-systems/service.ts
@@ -1,11 +1,7 @@
 "use server";
-import { Bundle, OperationOutcome, Parameters, ValueSet } from "fhir/r4";
-import {
-  ersdToDibbsConceptMap,
-  MISSING_API_KEY_LITERAL,
-} from "../../constants";
+import { Bundle, OperationOutcome, Parameters } from "fhir/r4";
+import { MISSING_API_KEY_LITERAL } from "../../constants";
 import { encode } from "base-64";
-import { indexErsdByOid } from "../db-creation/lib";
 import {} from "../db-creation/service";
 
 export type ErsdOrVsacResponse = Bundle | Parameters | OperationOutcome;
@@ -22,58 +18,6 @@ export type OidData = {
   oidToErsdType: Map<string, string>;
   conditions: Array<ersdCondition>;
 };
-
-/**
- * Performs eRSD querying and parsing to generate OIDs and extract clinical
- * concepts from the eRSD. First, a call is made to the eRSD, which is filtered
- * for valuesets. These valusets are used to create a mapping between OID and
- * clinical service code, as well as to actually compile the list of OIDs of all
- * valuesets in the pull-down. Then, these two entities are bundled together
- * as a data structure for future scripting.
- * @returns An OidData object containing the OIDs extracted as well as the
- * clinical service type associated with each.
- */
-export async function getOidsFromErsd() {
-  console.log("Fetching and parsing eRSD.");
-  const ersd = await getERSD();
-  const valuesets = (ersd as unknown as Bundle)["entry"]?.filter(
-    (e) => e.resource?.resourceType === "ValueSet",
-  );
-
-  const { nonUmbrellaValueSets, oidToErsdType } = indexErsdByOid(valuesets);
-  // Build up a mapping of OIDs to eRSD clinical types
-
-  let conditionExtractor: Array<ersdCondition> = [];
-  nonUmbrellaValueSets.reduce((acc: Array<ersdCondition>, vs: ValueSet) => {
-    const conditionSchemes = vs.useContext?.filter(
-      (context) =>
-        !(context.valueCodeableConcept?.coding || [])[0].system?.includes(
-          "us-ph-usage-context",
-        ),
-    );
-    (conditionSchemes || []).forEach((usc) => {
-      const ersdCond: ersdCondition = {
-        code: (usc.valueCodeableConcept?.coding || [])[0].code || "",
-        system: (usc.valueCodeableConcept?.coding || [])[0].system || "",
-        text: usc.valueCodeableConcept?.text || "",
-        valueset_id: vs.id || "",
-      };
-      conditionExtractor.push(ersdCond);
-    });
-    return conditionExtractor;
-  }, conditionExtractor);
-
-  // Make sure to take out the umbrella value sets from the ones we try to insert
-  let oids = valuesets?.map((vs) => vs.resource?.id);
-  oids = oids?.filter(
-    (oid) => !Object.keys(ersdToDibbsConceptMap).includes(oid || ""),
-  );
-  return {
-    oids: oids,
-    oidToErsdType: oidToErsdType,
-    conditions: conditionExtractor,
-  } as OidData;
-}
 
 /**
  * Fetches the eRSD Specification from the eRSD API. This function requires an API key
@@ -176,30 +120,4 @@ export async function getVSACValueSet(
       ],
     } as OperationOutcome;
   }
-}
-
-/**
- * helper function to generate VSAC promises
- *
- * @param oidsToFetch - OIDs from the eRSD to query from VSAC
- * @returns Promises to resolve that will give you the requested VSAC valuesets
- */
-export async function generateBatchVsacPromises(oidsToFetch: string[]) {
-  const valueSetPromises = Promise.allSettled(
-    oidsToFetch.map(async (oid) => {
-      try {
-        // First, we'll pull the value set from VSAC and map it to our representation
-        const vs = await getVSACValueSet(oid);
-        return { vs: vs, oid: oid };
-      } catch (e) {
-        let message = `Fetch for VSAC value set with oid ${oid} failed`;
-        if (e instanceof Error) {
-          message = message + ` with error message: ${e.message}`;
-        }
-        return Promise.reject(new Error(message));
-      }
-    }),
-  );
-
-  return valueSetPromises;
 }
