+++ src/app/tests/integration/decorators.test.ts
@@ -1,11 +1,8 @@
-import { dontUseOutsideConfigOrTests_getDbClient } from "@/app/backend/db/config";
 import { suppressConsoleLogs } from "./fixtures";
 import { insertFhirServer } from "@/app/backend/fhir-servers/service";
+import dbService from "@/app/backend/db/service";
 
-const dbClientQuery = jest.spyOn(
-  dontUseOutsideConfigOrTests_getDbClient(),
-  "query",
-);
+const dbClientQuery = jest.spyOn(dbService, "query");
 suppressConsoleLogs();
 
 describe("transaction decorator", () => {
